Refactor ResourceFetcher so that Config can accumulate errors about bad YAML files, etc.
